cleanup: eliminate unbounded/magic-sized sprintf calls.
I didn't spot anything that was obviously wrong but now it should all
be obviously correct.

Note that this introduces a dependency on vasprintf().  This function
can be found in Glibc and the BSDs (including OSX), but not AFAIK more
widely, and it has not been standardized despite its obvious utility.

If we need to support other targets then it's not too hard to
reconstruct it from vsnprintf().

Signed-off-by: Richard Kettlewell <rjk@terraraq.org.uk>